+++ src/frontend/web_application/src/components/Button/index.spec.jsx
@@ -1,16 +1,16 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { mount } from 'enzyme';
 import Button from './';
 
 describe('component Button', () => {
   it('should render', () => {
     const handleClick = jest.fn();
 
-    const comp = shallow(
+    const comp = mount(
       <Button onClick={handleClick}>Foo</Button>
     );
 
-    expect(comp.dive().text()).toEqual('Foo');
+    expect(comp.text()).toEqual('Foo');
     comp.simulate('click');
     expect(handleClick).toHaveBeenCalled();
   });