+++ src/frontend/web_application/src/modules/discussion/index.js
@@ -1,4 +1,5 @@
 export * from './actions/requestDiscussionIdForParticipants';
+export * from './actions/requestParticipantsForDiscussionId';
 export * from './selectors/discussionDraftSelector';
 export * from './selectors/discussionIdSelector';
 export * from './selectors/discussionSelector';