+++ src/frontend/web_application/src/scenes/Discussion/components/AddParticipantsToContactBook/index.jsx
@@ -1,10 +1,12 @@
-import React, { Component, Fragment } from 'react';
+import React, { Component, Fragment, createRef } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { withI18n } from '@lingui/react';
-import { ActionBarButton, Link, Icon, Dropdown, withDropdownControl, VerticalMenu, VerticalMenuItem } from '../../../../components';
+import { ActionBarButton, Link, Icon, Dropdown, withDropdownControl, VerticalMenu, VerticalMenuItem, TextBlock } from '../../../../components';
 import { getParticipantsExceptUser } from '../../../../services/message';
 import { withUser } from '../../../../modules/user';
+
+import './add-participants-dropdown.scss';
 
 const DropdownControl = withDropdownControl(ActionBarButton);
 
@@ -35,8 +37,10 @@ class AddParticipantsToContactBook extends Component {
       .filter(participant => !participant.contact_ids);
   }
 
+  dropdownControlRef = createRef();
+
   render() {
-    const { message, i18n, className } = this.props;
+    const { i18n, className } = this.props;
     const participants = this.getUnknownParticipants();
 
     if (participants.length === 0) {
@@ -46,15 +50,19 @@ class AddParticipantsToContactBook extends Component {
     return (
       <Fragment>
         <DropdownControl
+          ref={this.dropdownControlRef}
           display="inline"
           noDecoration
           className={classnames(className)}
-          toggleId={`message_${message.message_id}`}
           title={i18n._('message.action.add-participant-to-contacts', null, { defaults: 'Add a participant to the contact book' })}
         >
           <Icon type="address-book" /> <Icon type="plus" />
         </DropdownControl>
-        <Dropdown isMenu id={`message_${message.message_id}`}>
+        <Dropdown
+          dropdownControlRef={this.dropdownControlRef}
+          isMenu
+          className="m-add-participants-dropdown"
+        >
           <VerticalMenu>
             {participants.map(participant => (
               <VerticalMenuItem
@@ -67,7 +75,7 @@ class AddParticipantsToContactBook extends Component {
                   to={`/contact-association/${participant.protocol}/${participant.address}?label=${participant.label}`}
                   title={i18n._('message.action.add-to-contacts', null, { defaults: 'Add to contact book' })}
                 >
-                  {participant.label}
+                  <TextBlock inline className="m-add-participants-dropdown__label">{participant.label}</TextBlock>
                   {' '}
                   <Icon type="address-book" /> <Icon type="plus" />
                 </Link>